Added parameters to PI setup dialog
Added 2 Integer parameters for minMass and maxMass to the Peak
Investigator dialog via the PeakInvestigatorParameters class.
NOTE:  There is a bug in the IntegerParameter or
MassDetectionSetupDialog classes as the value of all Integer parameters
is set to 0.  There are no other examples of this working in MZMine so I
dont think this has even been tested before.